move flask tests into the main tests folder
